separates past and future events
