+++ client/src/scss/Home2.scss
@@ -20,7 +20,7 @@
   .left-pane {
     padding: 1vh;
     height: 90vh;
-    width: 65%;
+    width: 67.5%;
     @media (max-width: 800px) {
       width: 100%;
       height: 60vh;
@@ -29,7 +29,7 @@
   }
   .right-pane {
     overflow-y: scroll;
-    width: 35%;
+    width: 32.5%;
     height: 92.9vh;
     padding-right: 3px;
     // padding-left: 3px;
@@ -44,10 +44,11 @@
     header {
       display: flex;
       justify-content: space-between;
-      height: 40px;
+      height: 30px;
       border-bottom: 1px solid black;
       align-items: center;
-      padding: 2px;
+      padding: 5px;
+      padding-left: 0;
       padding-left: 2px;
       border-radius: 6px 6px 0 0;
       background: #ffffffb9;
@@ -97,12 +98,13 @@
       }
     }
 
-    .right-top {
+    .right-top,
+    .right-middle {
       box-shadow: 0px 0px 3px 2px rgb(235, 235, 235);
       border-radius: 6px;
       margin: 5px;
       margin-top: 1vh;
-      min-height: 420px;
+      min-height: 400px;
 
       h2 {
         background: rgb(217, 49, 37);
@@ -141,9 +143,17 @@
           width: 100%;
         }
         .loader {
-          padding-top: 30%;
+          padding-top: 20%;
           padding-left: 45%;
         }
+      }
+    }
+    .right-top {
+      margin-top: 1.5vh;
+    }
+    .right-middle {
+      h2 {
+        background: rgb(217, 49, 37);
       }
     }
     .right-compare {
