+++ client/src/views/App/Home2.js
@@ -4,6 +4,7 @@ import Map from "../../components/HomeView/Map/Map.js";
 import "../../scss/Home2.scss";
 import { connect } from "react-redux";
 import Charts from "../../components/HomeView/Charts.js";
+import Charts2 from "../../components/HomeView/Charts2.js";
 import Compare2 from "../../components/HomeView/Compare/Compare2.js";
 import Compare3 from "../../components/HomeView/Compare/Compare3.js";
 import Footer from "../../components/HomeView/Footer.js";
@@ -28,10 +29,15 @@ class Home2 extends Component {
           <div className="left-pane">
             <Map />
           </div>
+
           <div className="right-pane">
+            <div className="right-middle">
+              <Charts2 />
+            </div>
             <div className="right-top">
               <Charts />
             </div>
+
             <div className="right-bottom">
               <Compare2 />
             </div>
